render out gameController
