+++ js/screens/game.js
@@ -4,15 +4,9 @@
 
 import {Game1View, Game2View, Game3View} from './game-view';
 import {ANSWER_VARIETY, answers} from '../data';
+import {showNextGame} from '../gameController';
 
-import headerBlock, {startTimer, stopTimer, renderHeader} from './blockHeader';
-import statsBlock from './blockStats';
-import {levelTypes, GameType, headerData, statsData, currentLevel, MAX_LEVELS_AMOUNT} from '../data';
-import showNextScreen, {appendScreenElements} from '../utils/showNextScreen';
-import statsScreen from './stats';
-
-
-const game1 = () => {
+export const game1 = () => {
   const game1BlockView = new Game1View();
   game1BlockView.answerHandler = (event, answerType) => {
     const isCorrect = answers.check(event.target.value, answerType);
@@ -22,7 +16,7 @@ const game1 = () => {
   return game1BlockView.element;
 };
 
-const game2 = () => {
+export const game2 = () => {
   const game2BlockView = new Game2View();
 
   let answer1 = ``;
@@ -69,7 +63,7 @@ const game2 = () => {
   return game2BlockView.element;
 };
 
-const game3 = () => {
+export const game3 = () => {
   const game3BlockView = new Game3View();
   game3BlockView.answerHandler = (event, imageData, answerType) => {
     const selectedAnswer = imageData[event.target.id.slice(-1)][1];
@@ -78,51 +72,4 @@ const game3 = () => {
     showNextGame();
   };
   return game3BlockView.element;
-};
-
-
-const headerElement = headerBlock();
-let gameElement;
-let statsElement;
-
-export const resetAndStartGame = () => {
-  currentLevel.reset();
-  headerData.resetLives();
-  answers.reset();
-  levelTypes.reset();
-  statsData.reset();
-  showNextGame();
-};
-
-export const showNextGame = () => {
-  stopTimer();
-  if (currentLevel.level === MAX_LEVELS_AMOUNT || headerData.lives === 0) {
-    showNextScreen(statsScreen());
-    return;
-  }
-
-  currentLevel.up();
-
-  headerData.resetTime();
-
-  switch (levelTypes.levelsArray[currentLevel.level - 1]) {
-    case GameType.ONE_IMAGE:
-      gameElement = game1();
-      break;
-    case GameType.TWO_IMAGE:
-      gameElement = game2();
-      break;
-    case GameType.THREE_IMAGE:
-      gameElement = game3();
-      break;
-    default:
-    // default action
-  }
-
-  statsElement = statsBlock();
-
-  appendScreenElements(headerElement, gameElement, statsElement);
-  renderHeader();
-
-  startTimer();
 };
